ci: re-enable react-native-test-app integration test using local pack
Replace Verdaccio with local `yarn pack` + tarball install to fix
the hanging issue (#2344). Follows the same pattern already used by
the react-native-macos-init integration test.

Also moves and converts .ado/scripts/export-versions.mjs to
.github/scripts/export-versions.mts with GitHub Actions output support.
