+++ .github/workflows/microsoft-pr.yml
@@ -149,13 +149,11 @@ jobs:
     if: ${{ endsWith(github.base_ref, '-stable') }}
     uses: ./.github/workflows/microsoft-test-react-native-macos-init.yml
 
-  # https://github.com/microsoft/react-native-macos/issues/2344
-  # Disable these tests because verdaccio hangs
-  # react-native-test-app-integration:
-  #   name: "Test react-native-test-app integration"
-  #   permissions: {}
-  #   if: ${{ endsWith(github.base_ref, '-stable') }}
-  #   uses: ./.github/workflows/microsoft-react-native-test-app-integration.yml
+  react-native-test-app-integration:
+    name: "Test react-native-test-app integration"
+    permissions: {}
+    if: ${{ endsWith(github.base_ref, '-stable') }}
+    uses: ./.github/workflows/microsoft-react-native-test-app-integration.yml
 
   PR:
     name: "PR"
@@ -171,7 +169,7 @@ jobs:
       - build-rntester
       - prebuild-macos-core
       - test-react-native-macos-init
-      # - react-native-test-app-integration
+      - react-native-test-app-integration
     steps:
       - name: Check for failures or cancellations
         if: ${{ contains(needs.*.result, 'failure') || contains(needs.*.result, 'cancelled') }}
