+++ .github/workflows/microsoft-react-native-test-app-integration.yml
@@ -25,43 +25,38 @@ jobs:
         run: yarn install
 
       - name: Build community CLI plugin
-        run: yarn build
+        run: |
+          yarn build
+          # yarn build rewrites package.json exports; restore them
+          git checkout -- packages/*/package.json
 
       - name: Build react-native-macos-init
         working-directory: packages/react-native-macos-init
         run: yarn build
 
-      - name: Start Verdaccio server
+      - name: Export versions
+        id: versions
+        run: node .github/scripts/export-versions.mts
+
+      - name: Pack local react-native-macos
+        working-directory: packages/react-native
         run: |
-          set -euo pipefail
-          nohup npx --yes verdaccio --config .ado/verdaccio/config.yaml >/dev/null 2>&1 &
-          echo $! > $RUNNER_TEMP/verdaccio.pid
-      - name: Wait for Verdaccio to be ready
-        run: node .ado/scripts/waitForVerdaccio.mjs http://localhost:4873
-
-      - name: Configure npm for Verdaccio
-        run: .ado/scripts/verdaccio.sh init
-
-      - name: Publish to Verdaccio
-        run: .ado/scripts/verdaccio.sh publish --branch origin/${{ github.base_ref }}
+          set -eox pipefail
+          yarn pack -o ${{ runner.temp }}/react-native-macos.tgz
 
       - name: Clone react-native-test-app
         run: |
           git clone --filter=blob:none --progress https://github.com/microsoft/react-native-test-app.git
 
-      - name: Export versions
-        run: node .ado/scripts/export-versions.mjs
-
       - name: Configure react-native-test-app dependencies
         working-directory: react-native-test-app
         run: |
-          npm run set-react-version $(cat ${{ github.workspace }}/.react_native_version) -- --overrides '{ "react-native-macos": "1000.0.0" }'
+          npm run set-react-version ${{ steps.versions.outputs.react_native_version }} -- --overrides '{ "react-native-macos": "file:${{ runner.temp }}/react-native-macos.tgz" }'
 
       - name: Install dependencies in test app
         working-directory: react-native-test-app
         run: |
           set -eo pipefail
-          ${{ github.workspace }}/.ado/scripts/verdaccio.sh configure
           yarn --no-immutable
 
       - name: Bundle JavaScript
@@ -79,10 +74,3 @@ jobs:
         working-directory: react-native-test-app/example
         run: |
           ../scripts/build/xcodebuild.sh macos/Example.xcworkspace build
-
-      - name: Stop Verdaccio
-        if: always()
-        run: |
-          if [ -f "$RUNNER_TEMP/verdaccio.pid" ]; then
-            kill "$(cat $RUNNER_TEMP/verdaccio.pid)" || true
-          fi
